# Adjoint for diagonal should be lazy