Currently, `adjoint` for a `Diagonal` materializes the diagonal vector. This goes against the documentation of `adjoint` as a lazy wrapper.

This PR makes the result share memory with the original `Diagonal`.
```julia
julia> D = Diagonal(fill(2im, 2))
2×2 Diagonal{Complex{Int64}, Vector{Complex{Int64}}}:
 0+2im    ⋅  
   ⋅    0+2im

julia> D'
2×2 Diagonal{Complex{Int64}, Base.ReshapedArray{Complex{Int64}, 1, Adjoint{Complex{Int64}, Vector{Complex{Int64}}}, Tuple{}}}:
 0-2im    ⋅  
   ⋅    0-2im
```

 The implementation isn't perfect, as 
```julia
julia> D = Diagonal(fill(2im, 2))
2×2 Diagonal{Complex{Int64}, Vector{Complex{Int64}}}:
 0+2im    ⋅  
   ⋅    0+2im

julia> D'
2×2 Diagonal{Complex{Int64}, Base.ReshapedArray{Complex{Int64}, 1, Adjoint{Complex{Int64}, Vector{Complex{Int64}}}, Tuple{}}}:
 0-2im    ⋅  
   ⋅    0-2im

julia> (D')'
2×2 Diagonal{Complex{Int64}, Base.ReshapedArray{Complex{Int64}, 1, Adjoint{Complex{Int64}, Base.ReshapedArray{Complex{Int64}, 1, Adjoint{Complex{Int64}, Vector{Complex{Int64}}}, Tuple{}}}, Tuple{}}}:
 0+2im    ⋅  
   ⋅    0+2im
```
and, ideally, `(D')'` would just return `D`. However, at least this makes the `adjoint` operation lazy and allocation-free.