@dkarrasch 2023-01-31
The other option would be, of course, to make `D' = Adjoint(D)`, but then we would need to add a couple of multiplication and solve methods (which I removed at some point because one would not get an `Adjoint` via `adjoint`, irrespective of the `eltype`).