@jishnub 2023-04-18
Performance for some simple operations seems almost unaffected. If anything, solve seems slightly improved:
On master
```julia
julia> D = Diagonal([1:1000;]*im)';

julia> v = rand(ComplexF64, 1000);

julia> @btime $D * $D;
  1.611 μs (1 allocation: 15.75 KiB)

julia> @btime $D * $v;
  1.672 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $v;
  31.136 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $D;
  23.815 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $(Matrix(D));
  22.762 ms (2 allocations: 15.26 MiB)
```
This PR
```julia
julia> @btime $D * $D;
  1.610 μs (1 allocation: 15.75 KiB)

julia> @btime $D * $v;
  1.766 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $v;
  25.867 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $D;
  7.162 μs (1 allocation: 15.75 KiB)

julia> @btime $D \ $(Matrix(D));
  22.506 ms (2 allocations: 15.26 MiB)
```