@dkarrasch 2023-04-19
For the double adjoint, couldn't you add a method for `::Diagonal{<:Any,<:ReshapedArray{<:Adjoint}}` and make it do the ideal thing? And mitigate the issue further via

```julia
adjoint(D::Diagonal{<:Real}) = transpose(D)
```
? Then there should remain a smaller niche for this "ugly" behavior.